[#62] refactor: 디렉터리 이름을 handler에서 template으로 변경
